# server/swagger.yaml
openapi: 3.0.3
info:
  title: TochkaLi API
  description: |
    API документация для приложения TochkaLi
    
    ## Структура API
    
    API разделен на две части:
    - **Client** - эндпоинты для клиентского приложения (Telegram Web App)
    - **Admin** - эндпоинты для админ панели
    
    ⚠️ **Примечание:** authMiddleware и requireAdmin НЕ используются, все эндпоинты публичные.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API сервер
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    UserProduct:
      type: object
      properties:
        productId:
          type: string
        type:
          type: string
          enum: [subscription, one-time, free]
        paymentDate:
          type: string
          format: date-time
          nullable: true
        paymentAmount:
          type: number
        invoiceId:
          type: string
          nullable: true
        paymentStatus:
          type: string
          enum: [pending, paid, failed]

    # ==================== FAQ схемы ====================
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации
paths:
...
  /user/create:
    post:
      tags: [Client - Auth]
      summary: Создать пользователя (Telegram)
      description: Создает нового пользователя при первом входе через Telegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - telegramId
              properties:
                telegramId:
                  type: string
                  example: "123456789"
                telegramUserName:
                  type: string
                  example: "username"
                referralTelegramId:
                  type: string
                  description: Telegram ID пользователя, который пригласил
      responses:
        '200':
          description: Пользователь создан или уже существует
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
  /user/register:
    post:
      tags: [Client - Auth]
      summary: Регистрация пользователя
      description: Регистрация с ФИО, email и телефоном. Начисляет 10 бонусов.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - mail
                - phone
              properties:
                fullName:
                  type: string
                  example: "Иван Иванов"
                mail:
                  type: string
                  format: email
                phone:
                  type: string
                  example: "+7 999 999 99 99"
                telegramId:
                  type: string
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  userData:
                    $ref: '#/components/schemas/User'
        '400':
          description: Не все поля заполнены
  /user/login:
    post:
      tags: [Client - Auth]
      summary: Авторизация
      description: Вход по email и паролю (для админ панели)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  userData:
                    $ref: '#/components/schemas/User'
        '404':
          description: Неверный логин или пароль
  /user/send-mail:
    post:
      tags: [Client - Auth]
      summary: Отправить код подтверждения
      description: Отправляет 6-значный код на email. Cooldown 60 секунд.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mail
              properties:
                mail:
                  type: string
                  format: email
      responses:
        '200':
          description: Письмо отправлено
        '400':
          description: Пользователь с такой почтой уже существует
        '429':
          description: Повторная отправка возможна через N секунд
  /user/code-confirm:
    post:
      tags: [Client - Auth]
      summary: Подтвердить код
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mail
                - code
              properties:
                mail:
                  type: string
                  format: email
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Код подтвержден
        '400':
          description: Неверный код
  /user/send-mail-recovery:
    post:
      tags: [Client - Auth]
      summary: Код для восстановления пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mail
              properties:
                mail:
                  type: string
                  format: email
      responses:
        '200':
          description: Письмо отправлено
        '429':
          description: Слишком частые запросы
  /user/telegram/{telegramId}:
    get:
      tags: [Client - Auth]
      summary: Получить пользователя по Telegram ID
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
...
  /user/logout:
    post:
      tags: [Client - Auth]
      summary: Выйти из аккаунта
      description: Завершает активную сессию и инвалидирует refresh token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Сессия завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - PROFILE
  # ============================================================
  /user/profile:
    post:
      tags: [Client - Profile]
      summary: Получить профиль пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Профиль (с populated invitedUser)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
  /users/{telegramId}:
    patch:
      tags: [Client - Profile]
      summary: Обновить по Telegram ID
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phone:
                  type: string
                locatedInRussia:
                  type: boolean
                notifyPermission:
                  type: boolean
                password:
                  type: string
      responses:
        '200':
          description: Обновлено
        '404':
          description: Не найден
  /user/profile/update:
    put:
      tags: [Client - Profile]
      summary: Обновить профиль
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: Профиль обновлен
...
  /user/purchase-content:
    post:
      tags: [Client - Profile]
      summary: Купить контент за бонусы
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - contentId
                - contentType
              properties:
                userId:
                  type: string
                contentId:
                  type: string
                contentType:
                  type: string
                  enum: [practice, meditation, video-lesson]
      responses:
        '200':
          description: Контент приобретен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      bonus:
                        type: number
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserProduct'
        '400':
          description: Недостаточно бонусов

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - CONTENT (GET only)
  # ============================================================
...
  /horoscope/current:
    get:
      tags: [Client - Content]
      summary: Текущий гороскоп
      responses:
        '200':
          description: Текущий гороскоп
        '404':
          description: Не найден
  /horoscope/{id}:
    get:
      tags: [Client - Content]
      summary: Гороскоп по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Гороскоп
        '404':
          description: Не найден
  /meditation:
    get:
      tags: [Client - Content]
      summary: Все медитации
      parameters:
        - name: accessType
          in: query
          schema:
            type: string
            enum: [free, paid, subscription, stars]
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentItem'
                  count:
                    type: number
  /meditation/{id}:
    get:
      tags: [Client - Content]
      summary: Медитация по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Медитация
        '404':
          description: Не найдена
  /practice:
    get:
      tags: [Client - Content]
      summary: Все практики
      parameters:
        - name: accessType
          in: query
          schema:
            type: string
            enum: [free, paid, subscription, stars]
      responses:
        '200':
          description: Список
  /practice/{id}:
    get:
      tags: [Client - Content]
      summary: Практика по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Практика
        '404':
          description: Не найдена
  /video-lesson:
    get:
      tags: [Client - Content]
      summary: Все видео уроки
      parameters:
        - name: accessType
          in: query
          schema:
            type: string
            enum: [free, paid, subscription, stars]
      responses:
        '200':
          description: Список
  /video-lesson/{id}:
    get:
      tags: [Client - Content]
      summary: Видео урок по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Видео урок
        '404':
          description: Не найден
  /dynamic-content/name/{name}:
    get:
      tags: [Client - Content]
      summary: Динамический контент по имени
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Контент

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - SCHEDULE
  # ============================================================
  /schedule:
    get:
      tags: [Client - Schedule]
      summary: Все события
      parameters:
        - name: startDate
          in: query
          description: DD.MM.YYYY
          schema:
            type: string
        - name: endDate
          in: query
          description: DD.MM.YYYY
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
                  count:
                    type: number
        '400':
          description: Неверный формат даты
  /schedule/{id}:
    get:
      tags: [Client - Schedule]
      summary: Событие по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Событие
        '404':
          description: Не найдено

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - TRANSIT
  # ============================================================
...
  /transit/current:
    get:
      tags: [Client - Transit]
      summary: Текущий транзит
      responses:
        '200':
          description: Транзит
        '404':
          description: Не найден
  /transit/{id}:
    get:
      tags: [Client - Transit]
      summary: Транзит по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Транзит
        '404':
          description: Не найден

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - FAQ
  # ============================================================
...
  /faq/{id}:
    get:
      tags: [Client - FAQ]
      summary: FAQ по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: FAQ
        '404':
          description: Не найден

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - INFO
  # ============================================================
...
  /diary:
    post:
      tags: [Client - Diary]
      summary: Создать запись
      description: Начисляет 1 бонус (+1 за uselessTask)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                discovery:
                  type: string
                achievement:
                  type: string
                gratitude:
                  type: string
                uselessTask:
                  type: boolean
      responses:
        '201':
          description: Создано
        '401':
          description: userId обязателен

    get:
      tags: [Admin - Diary]
      summary: Все записи дневника
      responses:
        '200':
          description: Список
...
  /robres:
    post:
      tags: [System - Robokassa]
      summary: Webhook оплаты
      description: ResultURL для Robokassa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Обработано